requirements11.txt
pandas
numpy
scikit-learn
xgboost
joblib
matplotlib
seaborn

statsmodels
prophet
tsfresh

dvc
apache-airflow==2.6.3

apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-docker
apache-airflow-providers-http
apache-airflow-providers-slack

requests

supabase
psycopg2-binary  
pandas
python-dotenv
tqdm
schedule
dagshub